Update mopidy.conf
I think the config of mopidy changed. To get my setup to work as expected, I had to split the [local] into [core] and [local]. Otherwise it was not finding saving the scan.  Also the \local needed to be removed because it was creating it anyways.
The change in the config of mopidy can be seen here https://docs.mopidy.com/en/latest/config/